refactor: organize tests in two packages
